backend/requirements.txt
fastapi
uvicorn
pillow
pydantic
requests
yfinance
apscheduler
prisma
python-telegram-bot
python-dotenv
httpx
